# docs/swagger.yaml
basePath: /v1
definitions:
...
  Road-To-Destination-BE_module_trip_model_response.TripResponse:
    properties:
      createdTime:
        type: string
      endTime:
        type: string
      id:
        format: uuid
        type: string
      memberLimit:
        example: 15
        type: integer
      myRole:
        example: leader
        type: string
      name:
        type: string
      note:
        type: string
      ownerId:
        format: uuid
        type: string
      startTime:
        type: string
      status:
        example: planning
        type: string
      totalDistance:
        type: number
      tripType:
        example: bronze
        type: string
      updatedTime:
        type: string
      visibility:
        type: boolean
    type: object
...
  request.CreateTripRequest:
    properties:
      endTime:
        type: string
      mainBranch:
        $ref: '#/definitions/response.PreviewLocationsResponse'
      memberUserIds:
        items:
          type: string
        minItems: 1
        type: array
      name:
        maxLength: 255
        minLength: 1
        type: string
      note:
        maxLength: 4000
        type: string
      startTime:
        type: string
      tripType:
        example: bronze
        type: string
      visibility:
        type: boolean
    required:
    - memberUserIds
    - name
    - tripType
    type: object
...
  request.PreviewLocationsRequest:
    properties:
      locationIds:
        items:
          type: string
        minItems: 2
        type: array
    required:
    - locationIds
    type: object
...
  request.UpdateTripMemberRequest:
    properties:
      nickname:
        maxLength: 64
        minLength: 1
        type: string
      role:
        example: admin
        type: string
    type: object
...
  response.ComputeBranchLeg:
    properties:
      distanceM:
        type: number
      durationS:
        type: number
      from:
        $ref: '#/definitions/response.ComputeBranchStop'
      polyline:
        type: string
      to:
        $ref: '#/definitions/response.ComputeBranchStop'
      vehicle:
        example: bike
        type: string
    type: object
...
  response.ComputeBranchStop:
    properties:
      destinationId:
        format: uuid
        type: string
      locationId:
        format: uuid
        type: string
      name:
        type: string
    type: object
...
  response.PreviewLocation:
    properties:
      lat:
        type: number
      lng:
        type: number
      locationId:
        format: uuid
        type: string
      name:
        type: string
    type: object
  response.PreviewLocationsResponse:
    properties:
      legs:
        items:
          $ref: '#/definitions/response.ComputeBranchLeg'
        type: array
      locations:
        items:
          $ref: '#/definitions/response.PreviewLocation'
        type: array
    type: object
...
  response.TripInvitationListResponse:
    properties:
      invitations:
        items:
          $ref: '#/definitions/response.TripMemberResponse'
        type: array
    type: object
...
  response.TripJoinRequestListResponse:
    properties:
      joinRequests:
        items:
          $ref: '#/definitions/response.TripMemberResponse'
        type: array
    type: object
  response.TripListResponse:
    properties:
      trips:
        items:
          $ref: '#/definitions/Road-To-Destination-BE_module_trip_model_response.TripResponse'
        type: array
    type: object
  response.TripMemberListResponse:
    properties:
      members:
        items:
          $ref: '#/definitions/response.TripMemberResponse'
        type: array
    type: object
  response.TripMemberResponse:
    properties:
      invitorName:
        type: string
      joinedAt:
        type: string
      nickname:
        type: string
      role:
        example: member
        type: string
      status:
        example: active
        type: string
      tripId:
        format: uuid
        type: string
      tripName:
        type: string
      userId:
        format: uuid
        type: string
      username:
        type: string
    type: object
  share.ErrorResponse:
    properties:
      errors:
        additionalProperties:
          type: string
        type: object
      message:
        example: invalid input
        type: string
      status:
        example: 400
        type: integer
    type: object
info:
  contact: {}
  description: 'Backend API for Road2D (team Vandra): collaborative multi-branch ride
    planning and group-ride safety in Vietnam. JWT auth, Goong Maps playground, trip
    graph models.'
  title: Road2D
  version: "1.0"
paths:
...
  /planning/preview:
    post:
      consumes:
      - application/json
      description: Compute the bike route along ordered location ids before a trip
        exists. Nothing is stored. Pass this response as mainBranch on POST /trips
        to open the trip on that main branch.
      parameters:
      - description: Ordered location ids
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.PreviewLocationsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PreviewLocationsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/share.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Preview locations
      tags:
      - planning
  /trips:
    get:
      description: Returns trips where the caller is an active member (leader, admin,
        or member). Each item includes myRole. Left/kicked seats are omitted.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.TripListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/share.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List my trips
      tags:
      - trips
    post:
      consumes:
      - application/json
      description: Create a planning trip. The caller becomes leader and must name
        at least one other user, seated as a member. Bronze caps the trip at 15 seats.
        Trip type and member limit cannot be changed later. Optional mainBranch is
        the response of POST /planning/preview and becomes the trip's initial main
        branch. A group chat is not created yet.
      parameters:
      - description: Name, trip type, members, optional note, times, visibility, and
          reviewed main branch
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.CreateTripRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/Road-To-Destination-BE_module_trip_model_response.TripResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/share.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create trip
      tags:
      - trips
...
  /trips/{tripId}/invitations:
    post:
      description: The invited user responds with ?action=accept|reject. Accept seats
        them as a member when the trip is not full.
      parameters:
      - description: Trip UUID
        format: uuid
        in: path
        name: tripId
        required: true
        type: string
      - description: accept or reject
        enum:
        - accept
        - reject
        in: query
        name: action
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.TripMemberResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/share.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Accept or decline a trip invitation
      tags:
      - trips
...
  /trips/{tripId}/join-requests:
    get:
      description: Leader or admin. Pending rows only.
      parameters:
      - description: Trip UUID
        format: uuid
        in: path
        name: tripId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.TripJoinRequestListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/share.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List trip join requests
      tags:
      - trips
  /trips/{tripId}/join-requests/{userId}:
    post:
      description: Leader or admin. ?action=accept seats the user; reject sets status
        rejected. Accept fails with 409 when the member limit is already full.
      parameters:
      - description: Trip UUID
        format: uuid
        in: path
        name: tripId
        required: true
        type: string
      - description: Applicant UUID
        format: uuid
        in: path
        name: userId
        required: true
        type: string
      - description: accept or reject
        enum:
        - accept
        - reject
        in: query
        name: action
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.TripMemberResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/share.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Approve or reject a trip join request
      tags:
      - trips
...
  /trips/{tripId}/members:
    get:
      description: Active members only. Pending invites and join requests are separate
        lists.
      parameters:
      - description: Trip UUID
        format: uuid
        in: path
        name: tripId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.TripMemberListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/share.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List trip members
      tags:
      - trips
  /trips/{tripId}/members/{userId}:
    delete:
      description: Leader or admin. Cannot kick yourself, the leader, or (as an admin)
        another admin. The row stays as kicked so the user can be invited again.
      parameters:
      - description: Trip UUID
        format: uuid
        in: path
        name: tripId
        required: true
        type: string
      - description: Member UUID
        format: uuid
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Member kicked
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/share.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Kick trip member
      tags:
      - trips
    patch:
      consumes:
      - application/json
      description: An active member may change only their own nickname. The leader
        may also set another member's role to admin or member. Leadership is not assigned
        here.
      parameters:
      - description: Trip UUID
        format: uuid
        in: path
        name: tripId
        required: true
        type: string
      - description: Member UUID
        format: uuid
        in: path
        name: userId
        required: true
        type: string
      - description: Nickname and/or role
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.UpdateTripMemberRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.TripMemberResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/share.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/share.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update trip member
      tags:
      - trips
...
  /trips/invitations:
    get:
      description: Invitations a leader or admin sent to the caller. Pending join
        requests are not included.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.TripInvitationListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/share.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List my trip invitations
      tags:
      - trips
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
  PlaygroundKey:
    description: Shared secret from GOONG_PLAYGROUND_KEY. Scalar → Authorize.
    in: header
    name: X-Playground-Key
    type: apiKey
swagger: "2.0"